Update the production online pixel DQM.
This completely disables Phase0 DQM there, but the data remains Phase0.
Switched to file input and lxplus golbaltag.
This can be used for online development and merged with the one in SiPixelPhase1Config/testfor production.
Also needs to be switched back to Streamer input and P5 cluster conditions for Productions, but these settings cannot be tested currently.