Properly implement automatic per-lumi saving.
Now we use a separate flag, and MEs with this flag will behave differently
to those with lumiFlag in harvesting: they are *not* reset by run.
A complication is that now we need to explicitly store the value of the
lumiFlag in the DQMIO file. This is done by re-purposing the Flag branch
to store the ME flags (it used to store the Tag, which, as far as I see,
is completely unused).