Some basic booking functionality.
Includes a first guess at the internal data structures, going with a std::map keyed by ME metadata holding std::shared_pointers to MEs for now.
The Key tuple does not contain the ROOT object which makes it safe to handle, and it can be const and value to not corrupt the map.
It does mean we duplicate all the metadata for the key and in the object, but that might be not such a big deal.
Correctness before performance for now.