WIP: implement more of toProduct()
Itwill need a partner enterLumi(run, lumi) or similar to set up some
global fields that can be used to set the run for booking, and to
revive/recycle the MEs preserved here.

However, there is currently a big conceptual problem with concurrent
lumis: in that case, each DQMStore needs to have its own MEs, which the
DQMStore needs to map to one of nCuncurrentLumis real objects.