WIP: Restructure ME data.
MonitorElement is now a light, smart-pointer'y thing pointing to MonitorElementData.
Everything, including the lock, is in MonitorElementData. unique_ptrs hiold things together.
Due to all the non-copyable stuff in the product now we need to use persistent=false, maybe we could implement sufficient constructurs (cloning stuff) later to allow serialization. But for now, it is better to make sure we don't clone anything by accident.

Does not compile yet, ME implementation needs to be fixed.