WIP: lots of harvesting logic changes.
It is not so much that the logic changed, rather that there where many flaws and bugs before.
This hopefully makes the invariants more clear and follows them better.
There are still bugs (products agian are deallocated before they get to edm?), but we should be closer to corect than before.