Fix DQM/HLTEvF
Lots of ConcurrentME residues to clean up here.

I introduced a temporary in booking whenever things like axis labels are
set, to make things const-correct: the object returned by the booking
call is non-const and allows calling these methods, while the pointer
that is kept around for filling is const to make sure all operations are
thread safe.

Actually, in the current implementation even the non-const operations
are thread-safe (formally) but only filling makes sense in a
multi-threaded context.