Keep the correct products in the DQMIO output.
Specifying MonitorElementCollection as a type seems to be not possible,
edm seems to like to call it edm::OwnVector<...>, whihc as I understand
is not a legal thing to put into the OutputCommands. But for now,
grabbing them by the label is an ok solution as well.

We can now run a RECO production job with a very small, real DQM
sequence (DQMOfflineL1TEgamma) and get DQMIO output.