Fix SiStripFolderOrganizer.
So we can run ALCA step and get the same DQMIO output.

The problem was that SiStripFolderOrganizer sets folders on the
edm::Service DQMStore, and then the using modules assume that theis
IBooker books in the right place. The solution is to pass down a pointer
to the corect IBooker.