Implement enterLumi/leaveLumi.
This is the hard part -- at least half of it. This logic creates clones
if and only if they are required.

The other half is the cleanup code to run after saving, that needs to
delete all clones but a prototype after saving.

Not entirely sure yet what to do with harvesting -- we will have global
and local MEs as well, but I am not sure yet if we should separate them
by module ID. get*() will probably just return global MEs.