Remove the dirname* optimization from DQMNet.
Since we can have MEs whithout data attached, we need to keep a copy of
the name somewhere: Either by using a map or inside the ME. Since
DQMNet::CoreObject already has the name, we also use it to store the
path.

Doing that with this optimization would be possible, but complicated, so
I just removed it with the minimum amount of changes.