Some random changes to survive 136.85 step3.
The main thing is to make sure that only Scope::RUN histograms are used
in dqmEndRun. The changes here are more than required, but still not
complete.

In the longer term, we may want to restrict dqmEndRun to
DQMOneEDAnalyzer and completely ban endJob in DQM*EDAnalyzer.
The manual setScope would not be required then, since DQMOneEDAnalyzer
would use RUN scope by default.

Alternatively, we *could* support endRun in edm::stream, running it only
on a single stream, but that is unnecessarily complicated.