Some fixes to make more DQMIO tests pass.
Involves bugfixes in DQMStore as well as the DQMRootSource.
The current cleanup logic is pretty much dommed to fail, and it is not
really possible to get DummyReadDQMStore (relying on legacy emulation
behavior) to work now.

We could fix DummyReadDQMStore to not rely on the
legacy-keep-all-pointers-avlive behavior, but actually we should keep
supporting that for whatever else needs it.