Add a reScope option also to the EDMtoMEConverter.
It is equivalent to the DQMRootSource in role.
The default config always sets it to JOB, since that is what *all* existing workflows expect.
The not-JOB option is only used for the unit tests.

An issue is that harvesting code now can no longer tell if an ME was saved per lumi or not.